+++ src/main/java/com/example/aucison_service/controller/MypageController.java
@@ -5,6 +5,7 @@ import com.example.aucison_service.dto.ApiResponse;
 import com.example.aucison_service.dto.mypage.RequestAddressDto;
 import com.example.aucison_service.dto.mypage.RequestMembersInfoDto;
 import com.example.aucison_service.dto.mypage.RequestUpdateAddressDto;
+import com.example.aucison_service.service.address.AddressService;
 import com.example.aucison_service.service.member.MemberDetails;
 import com.example.aucison_service.service.member.MypageService;
 import lombok.RequiredArgsConstructor;
@@ -18,6 +19,7 @@ import org.springframework.web.bind.annotation.*;
 public class MypageController {
 
     private final MypageService mypageService;
+    private final AddressService addressService;
 
     @GetMapping("/buy") //구매 내역 조회
     @PreAuthorize("isAuthenticated()")
@@ -47,46 +49,46 @@ public class MypageController {
         return ApiResponse.createSuccess(mypageService.getSellInfo(principal));
     }
 
-    @GetMapping("/address") // 회원 정보 조회(배송지 조회)
-    @PreAuthorize("isAuthenticated()")
-    public ApiResponse<?> getAddressInfo(@AuthenticationPrincipal MemberDetails principal) {
-//        String email = principal.getAttribute("email");
-        return ApiResponse.createSuccess(mypageService.getAddressInfo(principal));
-    }
+//    @GetMapping("/address") // 회원 정보 조회(배송지 조회)
+//    @PreAuthorize("isAuthenticated()")
+//    public ApiResponse<?> getAddressInfo(@AuthenticationPrincipal MemberDetails principal) {
+////        String email = principal.getAttribute("email");
+//        return ApiResponse.createSuccess(mypageService.getAddressInfo(principal));
+//    }
 
     // 배송지 등록
     /*
     이 코드 기준으로 코드 작성
      */
-    @PostMapping("/address")
-    @PreAuthorize("isAuthenticated()")
-    public ApiResponse<?> addAddress(@AuthenticationPrincipal MemberDetails principal,
-                                     @RequestBody RequestAddressDto requestAddressDto) {
-//        MembersEntity members = principal.getMember();
-        mypageService.addAddress(principal, requestAddressDto);
-        return ApiResponse.createSuccessWithNoData("배송지 등록 성공");
-    }
+//    @PostMapping("/address")
+//    @PreAuthorize("isAuthenticated()")
+//    public ApiResponse<?> addAddress(@AuthenticationPrincipal MemberDetails principal,
+//                                     @RequestBody RequestAddressDto requestAddressDto) {
+////        MembersEntity members = principal.getMember();
+//        addressService.addAddress(principal, requestAddressDto);
+//        return ApiResponse.createSuccessWithNoData("배송지 등록 성공");
+//    }
 
-    // 배송지 삭제
-    @DeleteMapping("/address")
-    @PreAuthorize("isAuthenticated()")
-    public ApiResponse<?> deleteAddress(@AuthenticationPrincipal MemberDetails principal,
-                                        @RequestParam String addrName) {
-//        String email = principal.getAttribute("email");
-        mypageService.deleteAddress(principal, addrName);
-        return ApiResponse.createSuccessWithNoData("배송지 삭제 성공");
-    }
+//    // 배송지 삭제
+//    @DeleteMapping("/address")
+//    @PreAuthorize("isAuthenticated()")
+//    public ApiResponse<?> deleteAddress(@AuthenticationPrincipal MemberDetails principal,
+//                                        @RequestParam String addrName) {
+////        String email = principal.getAttribute("email");
+//        mypageService.deleteAddress(principal, addrName);
+//        return ApiResponse.createSuccessWithNoData("배송지 삭제 성공");
+//    }
 
-    //배송지 수정
-    @PatchMapping("/address") // 배송지 수정
-    @PreAuthorize("isAuthenticated()")
-    public ApiResponse<?> updateAddress(@AuthenticationPrincipal MemberDetails principal,
-                                        @RequestParam String addrName,
-                                        @RequestBody RequestUpdateAddressDto requestUpdateAddressDto) {
-//        String email = principal.getAttribute("email");
-        mypageService.updateAddressByEmailAndAddrName(principal, addrName, requestUpdateAddressDto);
-        return ApiResponse.createSuccessWithNoData("배송지 수정 성공");
-    }
+//    //배송지 수정
+//    @PatchMapping("/address") // 배송지 수정
+//    @PreAuthorize("isAuthenticated()")
+//    public ApiResponse<?> updateAddress(@AuthenticationPrincipal MemberDetails principal,
+//                                        @RequestParam String addrName,
+//                                        @RequestBody RequestUpdateAddressDto requestUpdateAddressDto) {
+////        String email = principal.getAttribute("email");
+//        mypageService.updateAddressByEmailAndAddrName(principal, addrName, requestUpdateAddressDto);
+//        return ApiResponse.createSuccessWithNoData("배송지 수정 성공");
+//    }
 
     // 택배 팝업
